Allow iovecs with buffers having length over 32bit.
Did some quick tests in Eio and it fixes that case:
See https://github.com/ocaml-multicore/eio/issues/335

There are probably more spots (outside iovec) where a (s)size_t is wrongly
expressed as an Unsigned.UInt, ideally it should always be a Long.

+++ src/helpers.ml
@@ -50,7 +50,7 @@ struct
       let base = Ctypes.(bigarray_start array1) bigstring in
       let length = Bigarray.Array1.dim bigstring in
       Ctypes.setf iovec C.Types.Buf.base base;
-      Ctypes.setf iovec C.Types.Buf.len (Unsigned.UInt.of_int length)
+      Ctypes.setf iovec C.Types.Buf.len (Unsigned.ULong.of_int length)
     end;
     iovecs
 end
